app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["neojelll <timofejmitusin60892@gmail.com>"]
readme = "README.md"

packages = [
	{ include = "app" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
aiokafka = "^0.12.0"
loguru = "^0.7.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.1"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.25.2"
httpx = "^0.28.1"
coveralls = "^4.0.1"

[tool.poetry.scripts]
run-app = "app.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=app --cov-report term-missing"

[tool.coverage.run]
omit = [
	"app/main.py",
	"app/logger.py",
	"app/models.py",
	"app/schemas.py",
]

[tool.ruff]
target-version = "py312"
indent-width = 4
fix = true

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[lint]
ignore = ["E501"]
